Patch Added global interceptor Added logout function

+++ Nova/src/app/components/navbar/navbar.component.ts
@@ -7,6 +7,7 @@ import { Product } from 'src/app/interfaces/product';
 import { DataService } from 'src/app/services/data.service';
 import { Subscription } from 'rxjs';
 import { Router } from '@angular/router';
+import {AuthService} from "../../services/auth.service";
 
 
 @Component({
@@ -61,7 +62,11 @@ export class NavbarComponent implements OnInit, OnDestroy {
     this.showDropDown = !(this.searchForm.value.search === null || this.searchForm.value.search === '');
   }
 
-  constructor( private fb: FormBuilder, _productsService: ProductsService, private data: DataService, private router: Router) {
+  constructor( private fb: FormBuilder,
+               private _productsService: ProductsService,
+               private data: DataService,
+               private router: Router,
+               private auth: AuthService) {
     this.initForm()
     this.productsService = _productsService;
   }
@@ -110,10 +115,15 @@ export class NavbarComponent implements OnInit, OnDestroy {
   }
 
   logout(){
-    sessionStorage.clear();
-    // alert("You are now logged out")
-    console.log("logged out")
-    // this.router.navigate(['/'])
+    this.auth.logout()
+      .subscribe(resp => {
+        if (resp.body == 'Successful Logout')
+        {
+          sessionStorage.clear();
+          console.log("logged out")
+          // this.router.navigate(['/'])
+        }
+      })
   }
 }
